requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
azure-ai-agents==1.0.0
azure-ai-ml==1.27.1
azure-ai-projects==1.0.0b11
azure-ai-resources==1.0.0b9
azure-common==1.1.28
azure-core==1.34.0
azure-core-tracing-opentelemetry==1.0.0b12
azure-functions==1.23.0
azure-identity==1.23.0
azure-mgmt-core==1.5.0
azure-mgmt-resource==22.0.0
azure-monitor-opentelemetry==1.6.7
azure-monitor-opentelemetry-exporter==1.0.0b36
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
azure-search-documents==11.5.2
aiohttp==3.10.1
ipywidgets == 8.0.7
ipykernel==6.28.0
ipython==8.16.0
python-dotenv==1.0.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
cryptography==45.0.3
databricks-sdk==0.55.0
Deprecated==1.2.18
fastapi==0.115.12
fixedint==0.1.6
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
h11==0.16.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
marshmallow==3.26.1
mlflow-skinny==2.22.0
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
openpyxl==3.1.2
opencensus==0.11.4
openpyxl==3.1.2
openstacksdk==1.2.0
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-logging==0.1.1
opentelemetry-api==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-dbapi==0.54b1
opentelemetry-instrumentation-django==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-instrumentation-flask==0.54b1
opentelemetry-instrumentation-psycopg2==0.54b1
opentelemetry-instrumentation-requests==0.54b1
opentelemetry-instrumentation-urllib==0.54b1
opentelemetry-instrumentation-urllib3==0.54b1
opentelemetry-instrumentation-wsgi==0.54b1
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
packaging==24.2
proto-plus==1.26.1
protobuf==6.31.0
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydash==8.0.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.25.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
strictyaml==1.7.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.22.0
